Preparing to branch off again, to do another (very minor in this case) parametric overhaul. The main changes will be:
1. Include all necessary environment information in the MAPF itself

Specific minor changes include tightening the interfaces:
1. Agents for search should always be selected by a single integer
index--not a vector of ints or anything else--passed to
low_level_search!()
2. env will be included as an argument for more of the relevant
functions, guaranteeing correct dispatch.

I have not yet decided how if/how to change the conflict/constraint
lookup interface. For the time being, they are working and I will leave
them as is.