Changed get_conflict_index() so that it actually checks for conflicts (assuming that a robot were to wait in place) instead of the hacky states_match() call, which would fail when PIBT is extended to heterogeneous robots.
About to try swapping out PIBTReservationTable for Set{Int} in
PIBTCache. Done.

Also, both reservation tables (PIBTReservationTable and
ReservationTable) are working with AbstractMetaEnvs. I just verified
that I can switch PIBTCache to use ReservationTable instead of
PIBTReservationTable. All tests pass just fine, which indicates that it
works out of the box.

Currently working to let agents check if an existing reservation belongs
to them (in which case they are allowed to change it). This becomes a
little bit trickier when dealing with teams of agents, where any agent
in the team should be allowed to change a reservation belonging to any
other member of the team.